# Conflicts:
#	.gitignore
#	gral-core/src/main/java/de/erichseifert/gral/data/DataAccessor.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/AbstractLegend.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/AbstractPlot.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/BarPlot.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/PiePlot.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/Plot.java
#	gral-core/src/main/java/de/erichseifert/gral/plots/XYPlot.java
#	pom.xml

+++ gral-core/src/main/java/de/erichseifert/gral/plots/colors/ContinuousColorMapper.java
@@ -1,3 +1,24 @@
+/*
+ * GRAL: GRAphing Library for Java(R)
+ *
+ * (C) Copyright 2009-2019 Erich Seifert <dev[at]erichseifert.de>,
+ * Michael Seifert <mseifert[at]error-reports.org>
+ *
+ * This file is part of GRAL.
+ *
+ * GRAL is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GRAL is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with GRAL.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package de.erichseifert.gral.plots.colors;
 
 import java.awt.Paint;
@@ -32,20 +53,19 @@ public abstract class ContinuousColorMapper extends AbstractColorMapper<Double> 
 
 	@Override
 	protected Double applyMode(Double value, Double rangeMin, Double rangeMax) {
-		if (value.doubleValue() >= rangeMin.doubleValue() &&
-				value.doubleValue() <= rangeMax.doubleValue()) {
+		if (value >= rangeMin && value <= rangeMax) {
 			return value;
 		}
 		Mode mode = getMode();
 		if (mode == Mode.REPEAT) {
 			return MathUtils.limit(value, rangeMin, rangeMax);
 		} else if (mode == Mode.CIRCULAR) {
-			double range = rangeMax.doubleValue() - rangeMin.doubleValue();
-			double i = value.doubleValue()%range;
+			double range = rangeMax - rangeMin;
+			double i = value%range;
 			if (i < 0.0) {
 				i += range;
 			}
-			return i + rangeMin.doubleValue();
+			return i + rangeMin;
 		}
 		return null;
 	}
